move syntax config to `highlight.ts`
